+++ wp_com_e2e/src/main.rs
@@ -10,6 +10,7 @@ mod languages_tests;
 mod me_tests;
 mod products_tests;
 mod purchases_tests;
+mod site_plans_tests;
 mod sites_tests;
 mod stats_city_views_tests;
 mod stats_country_views_tests;
@@ -75,6 +76,7 @@ fn collect_tests(ctx: Arc<TestContext>) -> Vec<Trial> {
     tests.extend(products_tests::tests(Arc::clone(&ctx)));
     tests.extend(transactions_tests::tests(Arc::clone(&ctx)));
     tests.extend(purchases_tests::tests(Arc::clone(&ctx)));
+    tests.extend(site_plans_tests::tests(Arc::clone(&ctx)));
     tests.extend(wp_service_tests::tests(Arc::clone(&ctx)));
     tests
 }